pyproject.toml
[project]
name = "multistatehnn"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "bokeh>=3.7.3",
    "hydra-core>=1.3.2",
    "jax>=0.6.1",
    "jaxtyping>=0.3.2",
    "matplotlib>=3.10.3",
    "numpy>=2.3.0",
    "plotly>=6.1.2",
    "scikit-learn>=1.7.1",
    "torch>=2.7.1",
    "torchvision>=0.22.1",
]

[dependency-groups]
dev = [
    "ruff>=0.12.12",
]

[tool.ruff]
target-version = "py313"
line-length = 100

[tool.ruff.format]
docstring-code-format = true

[tool.ruff.lint]
select = ["ALL"]
ignore = [
    "D1",       # undocumented
    "D203",     # one blank line before class
    "D212",     # multi-line summary second line
    "D400",     # First line should end with a period
    "D415",     # missing-terminal-punctuation
    "RUF001",   # 全角記号など`ambiguous unicode character`も使いたいため
    "RUF002",   # 全角記号など`ambiguous unicode character`も使いたいため
    "RUF003",   # 全角記号など`ambiguous unicode character`も使いたいた
    "TD001",    # invalid todo tag
    "TD002",    # missing todo author
    "TD003",    # missing todo link
    "PD011",    # pandas use of dot values
    "ERA001",   # Found commented-out code
    "T201",     # print
    "F722",     # jaxtyping
    "S101",     # use of `assert` statement
]
unfixable = [
    "F401", # unused import
    "F841", # unused variable
]

logger-objects = ["src.library.logger.LOGGER"]

[tool.ruff.lint.pylint]
max-args = 6